完善程序，修复bug，加入椭圆曲线加密（未完成）

+++ QuqiCrypto/base64.hpp
@@ -38,7 +38,7 @@ public:
             // 结束加密操作
             int flen = 0;
             EVP_EncodeFinal(shared_ctx.get(), (unsigned char*)out.data() + mlen, &flen);
-            out.resize(mlen + flen);
+            out.resize(size_t(mlen + flen));
             return true;
         }
         else
@@ -63,7 +63,7 @@ public:
             {
                 return false;
             }
-            out.resize(mlen + flen);
+            out.resize(size_t(mlen + flen));
             return true;
         }
     }
